* commit '2c0971eeec76f5a4b0aee890f4a05d47020da2b6':
  AArch64: Use long for pointers in services

+++ services/core/java/com/android/server/input/InputApplicationHandle.java
@@ -26,7 +26,7 @@ public final class InputApplicationHandle {
     // Pointer to the native input application handle.
     // This field is lazily initialized via JNI.
     @SuppressWarnings("unused")
-    private int ptr;
+    private long ptr;
 
     // The window manager's application window token.
     public final Object appWindowToken;
